async emitter tests
