+++ pkg/emitter/emitter.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"sync"
 
 	"github.com/go-logr/logr"
 	"github.com/tgoodwin/sleeve/pkg/event"
@@ -178,122 +177,4 @@ func (d *DebugEmitter) LogOperation(ctx context.Context, e *event.Event) {
 func (d *DebugEmitter) LogObjectVersion(ctx context.Context, r snapshot.Record) {
 	d.fileEmitter.LogObjectVersion(ctx, r)
 	d.InMemoryEmitter.LogObjectVersion(ctx, r)
-}
-
-type QueueItem struct {
-	ItemType string // "operation" or "record"
-	Event    *event.Event
-	Record   snapshot.Record
-	// We're not passing context through the queue since:
-	// 1. Context might be canceled by the time the item is processed
-	// 2. The underlying operations are just for logging, not for actual API calls
-}
-
-// AsyncEmitter is an emitter that buffers events in a channel to be processed asynchronously
-type AsyncEmitter struct {
-	underlyingEmitter Emitter
-	queue             chan QueueItem
-	wg                sync.WaitGroup
-	shutdown          chan struct{}
-}
-
-// NewAsyncEmitter creates a new AsyncEmitter with the specified buffer size
-func NewAsyncEmitter(underlying Emitter, bufferSize int) *AsyncEmitter {
-	ae := &AsyncEmitter{
-		underlyingEmitter: underlying,
-		queue:             make(chan QueueItem, bufferSize),
-		shutdown:          make(chan struct{}),
-	}
-
-	// Start the consumer goroutine
-	ae.wg.Add(1)
-	go ae.processQueue()
-
-	return ae
-}
-
-// processQueue processes items from the queue until shutdown is signaled
-func (ae *AsyncEmitter) processQueue() {
-	defer ae.wg.Done()
-
-	// Use background context for processing queued items
-	// since the original context might be canceled by the time we process the item
-	backgroundCtx := context.Background()
-
-	for {
-		select {
-		case item := <-ae.queue:
-			// Process the item
-			if item.ItemType == "operation" {
-				ae.underlyingEmitter.LogOperation(backgroundCtx, item.Event)
-			} else if item.ItemType == "record" {
-				ae.underlyingEmitter.LogObjectVersion(backgroundCtx, item.Record)
-			}
-		case <-ae.shutdown:
-			// Drain the queue before exiting
-			ae.drainQueue()
-			return
-		}
-	}
-}
-
-// drainQueue processes all remaining items in the queue
-func (ae *AsyncEmitter) drainQueue() {
-	// Use background context for processing remaining queued items
-	backgroundCtx := context.Background()
-
-	// Keep processing until the queue is empty
-	for {
-		select {
-		case item := <-ae.queue:
-			if item.ItemType == "operation" {
-				ae.underlyingEmitter.LogOperation(backgroundCtx, item.Event)
-			} else if item.ItemType == "record" {
-				ae.underlyingEmitter.LogObjectVersion(backgroundCtx, item.Record)
-			}
-		default:
-			// Queue is empty
-			return
-		}
-	}
-}
-
-// LogOperation queues an operation event
-func (ae *AsyncEmitter) LogOperation(ctx context.Context, e *event.Event) {
-	select {
-	case ae.queue <- QueueItem{
-		ItemType: "operation",
-		Event:    e,
-	}:
-		// Item queued successfully
-	case <-ctx.Done():
-		// Context canceled, drop the event
-		fmt.Println("WARNING: Context canceled, dropping operation event")
-	default:
-		// Queue is full, log a warning and drop the event
-		fmt.Println("WARNING: AsyncEmitter queue is full, dropping operation event")
-	}
-}
-
-// LogObjectVersion queues an object version record
-func (ae *AsyncEmitter) LogObjectVersion(ctx context.Context, r snapshot.Record) {
-	select {
-	case ae.queue <- QueueItem{
-		ItemType: "record",
-		Record:   r,
-	}:
-		// Item queued successfully
-	case <-ctx.Done():
-		// Context canceled, drop the record
-		fmt.Println("WARNING: Context canceled, dropping object version record")
-	default:
-		// Queue is full, log a warning and drop the record
-		fmt.Println("WARNING: AsyncEmitter queue is full, dropping object version record")
-	}
-}
-
-// Shutdown gracefully shuts down the AsyncEmitter, ensuring all queued events are processed
-func (ae *AsyncEmitter) Shutdown() {
-	close(ae.shutdown)
-	ae.wg.Wait()
 }
